% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Create odin.api}
\usage{
api(validate = NULL)
}
\arguments{
\item{validate}{Logical, indicating if validation should be done
on responses.  This should be \code{FALSE} in production
environments.  See \link[porcelain:porcelain]{porcelain::porcelain} for details}
}
\value{
A \link[porcelain:porcelain]{porcelain::porcelain} object. Notably this does \emph{not}
start the server
}
\description{
Create an odin.api server, a porcelain object
}
