% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{server}
\alias{server}
\title{Run odin.api server}
\usage{
server(port, host = "0.0.0.0")
}
\arguments{
\item{port}{Port to serve on}

\item{host}{Optional host (either \verb{0.0.0.0} or \verb{127.0.0.1})}
}
\value{
Never returns
}
\description{
Run odin.api server
}
